Added memory logging using PetscLogObjectMemory. However, the underlying objects blow up upon destruction (I think because they have no type set), while leaving them undestroyed doesn't seem to log the memory.
Changed #ifdefs that control logging slightly (see makefiles).  In particular, ALE_LOGGING_VERBOSE will turn on a lot of
event registration output (why do we even do that?).  Perhaps best not to turn it on.

Note that the library must be rebuilt in dm/mesh.

Some minor changes to Map.hh and some Notes on validating (re)design of Sieve.


Hg-commit: 42e0659a108d5b47810523498f992913f93aada9