Added the Nash variant of the conjugate gradient method for minimizing a function subject to a trust-region constraint.  (This method does not follow directions of negative curvature.)
Update the stcg and gltr methods to count matrix-vector products for the
iteration count, force at least one iteration to be performed, fix
problems with the monitor, use the gradient direction if the first
direction computed has zero norm, and fix the ordering of the
convergence tests so that the last matrix-vector product is not
wasted.

The papers describing all the methods are now documented in the header
file.  These methods can be used for any problems; if the radius is
not set, then they mimic the standard conjugate gradient method with
a few minor differences.

Hg-commit: 9b0bf2eae71f5c9b599afbe7be9adec8ef05e21d