From:	msaum@math.utk.edu 	Subject:	Re: [PETSC #18960] Error Compiling Petsc with matlab 	Date: 		January 30, 2009 3:13:24 PM CST One thing you might consider on future revisions is to eliminate the use of the matlab command "flops" contained in src/sys/matlabengine/matlab.c.
In a recent version of matlab (7.6.0.324 (R2008a) on linux), I get the following:

>> flops(0)
??? Error using ==> flops at 10
FLOPS is obsolete and will be removed in a future version.

This also gives an error in petsc when run, but I was able to make appropriate temporary mods to petsc so that snes/examples/tutorials/ex5.c runs as it should to completion.

Hg-commit: 4ab7d275a93622b98b3628028817d42a3a467aa1