MatGetSubMatrix API change
The user now specifies a parallel IS for the columns.  The local part of
this IS corresponds to the local columns of the matrix.  Note that
MatPermute_MPIXAIJ calls the private versions where the full index set
is given.  It is probably desirable for MatPermute to also use a
parallel IS rather than require a gathered IS.

Ripped out MatGetSubMatrixRaw, it is not used anywhere in PETSc.

Hg-commit: 3b40e47b3df6b69dfffc2cc67b61dab74aafd570