asm.c: create osm->is_local when needed and handle nontrivial subdomains
Multiple subdomains per process used to only work when created specially
and when they happened to be partitioned into contiguous pieces (instead
of by a graph partitioner).  This patch uses ISGlobalToLocalMappingApply
to generate the scatter from the vector corresponding to each osm->is[i]
(includes overlap) to the vectors based on osm->is_local[i], instead of
assuming that each subdomain was contiguous.

Hg-commit: d63b278424668f58c027d41777bd75cfb6383bd7