added info on open-mx to faq.html
      From:   jed@59A2.org
      Subject:	[petsc-users] Open-MX: a low-latency alternative to IP over 1 and 10 Gigabit Ethernet
      Date: 	February 4, 2010 9:31:03 AM CST
      To: 	petsc-users@mcs.anl.gov
      Reply-To:		petsc-users@mcs.anl.gov
This may be interesting to those solving with relatively small
subdomains on ethernet clusters.  Open-MX is an open source
implementation of the Myrinet Express (MX) protocol which operates over
generic ethernet hardware.  It is a driver for Linux (>= 2.6.15) and
does not interfere with IP traffic.  It works with the native MX backend
of several MPI implementations including Open MPI and MPICH2.

The big feature for PETSc users is that compared to TCP/IP, it offers
30% lower latency over GigE and half the latency with 25-30% higher
bandwidth over 10 GigE.  This could make a big difference for those
observing poor scalability on ethernet clusters.  In particular, if
-log_summary shows a lot of time in events like VecDot, VecNorm, and
VecScatterEnd.

Hg-commit: c9084d8c0ddf9b17f87f3d996b23e1b8ef80e886