Update ML interface to support energy minimization and other options.
This uses many features that are not documented in the reference manual.
In particular, the C interface does not seem to be supported any more,
the current one is based on strings and is only available from
C++/Epetra with associated baggage and certain limitations relative to
the C interface.  Fortunately, the C++ interface has a nearly one-to-one
correspondence with fields in the ML_Aggregate structure, so the present
code sets options in that structure directly (when not available from
C++).

The biggest problem with exposing this extra functionality is that ML
options are rather non-orthogonal so it's easy to produce options sets
that are not actually functional.  Rather than providing useful
(user-understandable) information about what is wrong, ML calls exit(1).
So if ML crashes, you have to attach a debugger with a breakpoint on
'exit' just to get a stack trace (for example, to know which ML instance
in your code is causing trouble).  I am less than thrilled with this
error handling strategy.

Hg-commit: e721d2184f87ad8008519982dabd8955327e31f2