Move processing of -fp_trap to come after pushing other signal handlers
This prevents the trapping signal handler for SIGFPE from being overwritten
by the default handler which can't provide a useful error message.

Note that the lack of interoperability between PetscPushSignalHandler
and floating point handlers is a deficiency in the interface for the
former.  All the solutions I can think of would add nontrivial
complexity so I'm not implementing them now.

Hg-commit: 97a23cb8dc137879bc69e35ee6a7c964fcfff5ee