From:	Blaise Bourdin <bourdin@lsu.edu> 	Subject:       [petsc-maint #47679] fortran bindings for random seed and friends 	Date: 	       June 13, 2010 6:08:27 PM CDT
s there a reason why the automatic fortran bindings for PetscRandomSeed, PetscRandomGetSeed and PetscRandomSetSeed are not enabled?
I did enable them, and they work fine. Unless this is causing problems in other places or there is another way to seed a random number generator, would it be possible to enable these bindings by default in petsc-dev?

Hg-commit: d8fd437c9de9351fb7ace0eaa1b2c53787b97f25