Make PetscOptionsGetString ensure that the result is NULL terminated, update uses.
The former usage sometimes passed len-1 and sometimes len
(inconsistently).  Unfortunately, even using a shorter value does not
guarantee that the string is NULL terminated (because strncpy does not
terminate if the source string is longer than len).  Since both uses
were incorrect in the presence of overflow, this patch chooses to
specify the length of the buffer rather than the maximum number of
non-NULL characters.

Hg-commit: cc94bdc2460eb6696fbd1c57d30a0d1f633d10f4