Fix ISComplement when the input IS has duplicates.
The old version allocated nmax-n space, which is not sufficient for
nmin=0 with duplicate entries (memory corruption).  It is also way too
much space whan nmin is large (potential space bottleneck).  The new
version scans the array to find the number of unique values, then
allocates exactly the correct amount of space (nmax-nmin-unique) for the
complement, which is passed off with ISCreateGeneralNC.

Hg-commit: 5df35d84b28ba16e9d6913ebc9f54c9604e3f2d0