Add nlocal field to DMCompositeLink
Rationale: n was used for different purposes, owned part for DM, and
local part for Array.

Also changed DMCompositeGetGlobalISs to return the same number of ISs on
every process, instead of only returning for owned part of the Array
entries.  This function was guaranteed to deadlock before because only
one process was calling ISCreateGeneral(((PetscObject)dm)->comm,...),
but this call is necessarily collective.

Hg-commit: 4d55b5a7a67069ac539ee945f0eff1496ccfe0f9