Better fieldsplit handling for Stokes using MatNest and -pc_fieldsplit_real_diagonal
User provides a MatNest J = [A B; C 0] with a truly empty block, then
also provides Jp = [Ap 0; 0 D] where Dp is an approximation of the Schur
complement (e.g. scaled mass matrix). We want to use the real diagonal
to form the kspschur, then the diagonal block will be used for
preconditioning.

Hg-commit: 1433b73803a93eea43be2ff844c1beba570a75a4