Place MPI_Barrier in the necessary places for the STREAM benchmarks.
Note that it is possible for the system to interrupt the process between
the call to MPI_Barrier and getting the current time. This means that
scenarios like

  MPI_Barrier(comm);
  t1 = time()
  work()
  MPI_Barrier(comm);
  elapsed = time() - t1;

may have each work() running concurrently, or they may run separately,
paired with other active processes on the system. If the other processes
do not have similar memory demands, the aggregate throughput can be
reported as anomalously high. Therefore it is necessary to bracket the
computation with MPI_Barrier.

Hg-commit: def515087559f3ce819b0b673d0ceabaa58d5a03