Fix problem in TSGL where TSGLAdaptSetFromOptions() was not reached.
PetscOptionsTail() contains a return statement, so code cannot reliably
be placed after it (in light of the deeper issue below).

I need to remove TSGLAdapt and make TSGL use TSAdapt, which the other
methods are already using.

Deeper issue: PetscOptionsBegin() unconditionally increments
PetscOptionsPublishCount. This should be changed so that
PetscOptionsBegin() is like PetscOptionsHead() when called within an
active options loop. That is too big of a change for petsc-3.2, so it
does not accompany this patch.

Hg-commit: 415528188cad0cf25e81fc1091b06c0e443b94ab