Refactor SNES DM infrastructure
The residual and Jacobian function pointers are now stored in a managed
context on the DM. Coarsening of this context is handled by a
DMCoarsenHook, so the coarse level DM automatically has function
pointers for residual evaluation. The SNESDM context (attached to a DM)
has copy-on-write semantics for all DMs other than the one it was
originally created on. That means that changes to the function pointers
from the fine level are automatically propagated unless a coarse level
is changed manually.

Hg-commit: ce0066853e1ef4402603a33070cb3bfd75c87f86