Cache the threadcomm in PetscCommGetThreadComm
This cache is unsafe if it is possible for a given MPI_Comm value to be
reused as a different communicator. Need to check with implementations
if this is actually likely to happen.

The problem is that PetscCommGetThreadComm is somewhat expensive since
it also calls through PetscCommDuplicate() and PetscCommDestroy(). The
MPI_Attr_get() call is on the order of 100 cycles so one is not
catastrophic, but several, combined with other logic, is a concern. We
may want to cache the threadcomm on the outer comm to ensure that the
call is only made once.

Hg-commit: 85092c7f6e43b1ecdbc83cb68206d9f92846b45e