Add TSSetPreStage() and move call to TSPreStep() into stepping routines
TSPreStage(ts,stage_time) is a hook called before starting to evaluate a
stage.

The call to TSPreStep() has been moved into the TSStep_X implementations
because it is called again after a step is rejected. This allows prestep
hooks to do things like Monte-Carlo sampling over the time interval
(dependent on the length of the time step), the result of which will
appear as a continuous source term during the time step.

Hg-commit: c5d083c79bbad0900afc1d94d1490cb90cadff0c