ts ex11: split up initial conditions and verify order of accuracy for unlimited method
Tested with the smooth bump-in-annulus problem, using ssp-rk104 to
keep time integration error small. This method has a stability limit of
6, but since there is a thin buffer of constant solution away from the
maximum speed, we use the full stability limit.

./ex11 -ts_ssp_type rk104 -physics advect -advect_sol_type bump -advect_bump_type cos -ufv_cfl 6 -ts_final_time 2 -ufv_reconstruct 1 -ufv_limit none -monitor Error
-f annulus-10.exo:  34  time    2.015  |x|   0.9564         Error [1.046326e-25,0.07640565] int 0.09302936
-f annulus-05.exo:  81  time    2.015  |x|   0.9935         Error [2.080105e-48,0.02133695] int  0.0267694
-f annulus-03.exo: 123  time     2.01  |x|   0.9988         Error [1.115783e-75,0.009725027] int 0.01037744

Since CFL is set relative to the smallest element radius, we can use the
number of time steps as a proxy for inverse grid size on this
unstructured mesh. This gives the estimated convergence rates

(log(0.07640565) - log(0.02133695)) / (log(34) - log(81))  = -1.4695
(log(0.02133695) - log(0.00972502)) / (log(81) - log(123)) = -1.8809

Hopefully the convergence rates would be cleaner using a structured
mesh. I am seeing much worse convergence when using a limiter, but I
have not yet concluded whether it is an inherent problem with scalar
limiting or whether the limiting formulation is flawed, though I have
verified that arbitrary linear functions are reconstructed exactly.

Hg-commit: 6570af799159503c2067f341d1f47fd28f4b6ace