scalapack: update to 2.0.2
The new version of scalapack no longer depends on BLACS so we no longer need to specify that dependence to MUMPS.

Also, the genius that decided to use the case-sensitive spelling of a package's filename within buildsystem has ensured that there will *always* be future repository problems since most of the developers (e.g. Barry) have case-insensitive filesystems. This will be problematic for bisection as well.

Hg-commit: 6452035a0fe16cff7e7ff959c07b52f2708cf2a0