Fieldsplit cannot force its algorithmic block structure on solution vectors
This problem shows up in snes ex19_8 which does

  -pc_fieldsplit_block_size 2
  -pc_fieldsplit_0_fields 0,1
  -pc_fieldsplit_1_fields 0,1

even though the original system has data structure block
size 4. Although this particular set of options is contrived, I consider
it (and variants) to be valid and PCFieldSplit should work.

This reverts the change introduced in

  398f06830ac9d0920f8d4e4d0bac972857d520d4

and modified (thus activating the error message) in

  1b382c0c8a0048fa000c4641e30980e1ca800f19
  Make fieldsplit only attempt to set vector block size when jac->bs > 0

Hg-commit: 85479e58af594607b0eeb3032c0aea35cfb6567e