Add support for internal symbol visibility
PETSC_INTERN is for symbols that are shared between files, but only used
within one shared library. Note that PETSC_EXTERN must be used for
symbols that are accessed from other shared libraries (e.g., a symbol in
petscmat.so that is accessed from petscksp.so).

PETSC_INTERN_C is for functions with internal visibility that must use
the `extern "C"` calling convention. Normally this is any symbol passed
to a dynamic registration function.

Hg-commit: 8413f6e91807356123f19f648513e17d0ebc6e4e