CMake-CUDA: Circumvent problems of passing -arch=sm_XY to CMake
The problem is that the CMake parser has problems with additional '=' signs, such as in CUDA_NVCC_FLAGS:STRING="-g -arch=sm_13"
Could not find a consistent way of passing these flags, hence using an initial cache file for this.
Also increased timeout in order to avoid premature termination of sane CMake runs.