PetscFunctionListAdd: use macro to hide (PetscVoidFunction) cast
Same rationale as the last commit:

Casting using PetscVoidFunction instead of (void(*)(void)) is needed
with C++ compilers that will not implicitly convert C++ language linkage
to C language linkage. Since PetscFunctionListAdd() is meant to operate
on generic function pointers, we hide this language linkage detail
(which is inconsistently implemented) in a cast.