ksp tests: new tests for ksp algorithms using aijcusparse methods
A new test, ex43-aijcusparse.c, is added to the src/ksp/ksp/examples/tests
directory. This example tests various combinations of the aijcusparse class
including different storage formats and both the ilu and icc
preconditioners. Results are committed for the K20c architecture.
However tests on the 2070 architecture give identical behaviour as
expected. For ICC preconditioning (runex43-aijcusparse_2), this test now
runs a much smaller matrix (shallow_water1) and it converges fast. The
symmetric shallow_water1 matrix is stored in upper triangular form.
The aijcusparse class handles symmetry properly by informing the class
the matrix is symmetric as demonstrated in the test. New results are
being committed here.

Finally another target is added for the ex10 example in tutorials
subdirectory of the ksp examples. This target ensures the ex10 class
is test with the aijcusparse class when built with txpetscpgu.