missing.py: replace checkPrototype() with assignment to function pointer
C89 does not produce a warning when calling a function that has not been
declared, but if we assign the function to a function pointer, it is an
error for the function to have not been declared. This makes the checks
for missing declarations more strict and allows removal of the
checkPrototype() function.