aijcusparse: ICC/Cholesky preconditioners
This is a patch for ICC/Cholesky preconditioners on GPUs for the
aijcusparse(aijcusp) class. One uses ICC by loading an upper triangular
matrix into the cusparse class, and then setting the symmetry option,
i.e.

ierr = MatCreateSeqAIJCUSPARSE(comm,n,m,PETSC_NULL,num_entries_per_row,&A);CHKERRQ(ierr);
ierr = MatSetOption(A,MAT_SYMMETRIC,PETSC_TRUE);

This contains contains both the organization and implementation of the
preconditioner. The solves are done under the hood in the txpetscgpu
package. Currently, the factorization is done in the aij/seq/aijfact.c
routines. aijcusparse.cu makes scoping calls to these methods. Then, the
matrix is rebuilt into a form that is palatable for the GPU. This is
done in the method MatSeqAIJCUSPARSEBuildICCTriMatrices.

[Jed] Formatting and 'static' cleanup