Stack: consolidate handling of PetscStack
This is in preparation for addition of PetscStackCallBLAS(), for which I
would like to avoid further duplication (it has already led to
petscstackp->petscroutine becoming stale in case of PetscStackPush).

This change has only one functional change: PetscStackPush will now
report the file and line number of the call site instead of "unknown"
and 0.  This is more useful because it distinguishes multiple calls from
the same function.