*ViewFromOptions: update documentation, providing disabled man pages
Most uses for these functions are to place diagnostics in the main
execution flow rather than placing it in a normal *SetFromOptions
function that is only executed once.  This commit provides documentation
because the old comments were incorrect, but does not generate man pages
because we'd rather not encourage users to choose this interface.