config: add libraries.checkClassify to find missing functions, use in BlasLapack
Usually we find most of the functions that we need to look for, so be
optimistic and try to compile and link them all at once.  If that fails,
recursively bisect so that we usually perform less than one compile/link
per function that we need to check for.

Using this in BlasLapack cuts BlasLapack search time in half.