make: add GNU make and ninja build system
Currently only works for shared, single-library, and gcc-compatible
compilers.  The Fortran modules are a soup of circular dependencies,
which make warns loudly about.

  make -f gmakefile PETSC_ARCH=your-arch -j10

This requires GNU make version 3.81 or later.

For Ninja [1], the entire file is generated in the build directory, so run

  conf/gmakegen.py PETSC_ARCH=your-arch --output=ninja
  ninja -C $PETSC_ARCH

The Ninja build does not yet support C++.

[1] http://martine.github.io/ninja/