TS: fix shortcut scaling; document TSComputeIJacobianConstant limitations
The shortcut re-shifting on re-evaluation of the IJacobian can only be
used when the implicit system is just the "mass matrix".  This
re-scaling also needs to be recorded for the next iteration.  Since this
logic cannot work in general, it is now moved to
TSComputeIJacobianConstant.  This funtion seems to be of limited use
beyond "explicit" finite element computations (especially those that use
a consistent mass matrix) and similar.