TS: add TSRHSJacobianSetReuse, fixes TSComputeRHSJacobianConstant
When using a constant RHS and no IJacobian, the RHS Jacobian is shifted
to make it function as an IJacobian (shift*I - Arhs), which is the matrix
used by SNES.  We don't want to duplicate the full matrix because that
uses a lot of memory, so we play this game of recording the shift and
scale, then unshifting in TSComputeRHSJacobian.  Add tests that fail
prior to this bug fix.