Removing TXPETSCGPU from CUSP matrix and preconditioner classes
The default type of the matrix in Mat_SeqAIJCUSP struct is set
to a void *. Then it is cast to a particular cusp matrix type
(csr, ell, dia) depending on the chosen format by the user with
csr being default. This need to propagate into the the cusp
preconditioner classes.

A stream variable was added to the Mat_SeqAIJCUSP struct. This
is allocated and deleted by the mpiaijcusp class--in serial
the stream variable is set to 0. Ultimately, the CUSP API will
need to support streams in order to get an efficient multi-GPU
SpMV. This code for allocating, setting, deleting streams is
added so that once the CUSP API is enhanced to accept streams,
the interface will work naturally.