Fix to AIJCUSP to Enable larger parallel simulations with Ellpack Storage
The construction of the offdiagonal matrices has been changed to enforce
compressed row format. Before this commit, the entire CSR format of the
offdiagonal block matrices (used in MatMultAdd) was being used to convert
to ellpack. This lead to a huge amount of fill because many of the rows in
the offdiagonal blocks can often be empty. With this commit, I calculate
the compressed format and calculate the ellpack matrix from this form.
No change in test results.