Sys Logging: revert parent traversal
Traversing parents is dangerous because we can have dangling links.
The problem is explained in more detail on petsc-dev:

  Subject: Memory corruption on 'master' (barry/improve-memory-logging)
  Message-ID: <87ob8snnht.fsf@mcs.anl.gov>

but the pattern of creating an inner solver with the outer PC is an
example that leaves a dangling parent.

    KSPCreate(PETSC_COMM_WORLD,&eksp);
    KSPGetPC(ksp,&pc);
    KSPSetPC(eksp,pc);
    // Use eksp for something, or not
    KSPDestroy(&eksp);  // pc->parent still points at eksp

In the general case, objects could have multiple parents and multiple
children.  How would we avoid over-counting?  How do we make the
addition or removal of parent-child relationships commutative in the
sense that the memory logging would be the same?