Remove --with-dynamic-loading / PETSC_USE_DYNAMIC_LIBRARIES
This option was mostly redundant because normal shared libraries satisfy
all the same needs.  This commit adds the run-time option
-dynamic_library_preload that can be used to load/initialize all the
packages in advance, as used to be done when --with-dynamic-loading.

The PetscDLLibraryRegister_petscxxx functions have been kept, but are
not called unless -dynamic_library_preload is used.  The petsc-dev
discussion, "Why --with-dynamic-loading?" did not reach consensus about
whether this was ever useful, so it may also be removable.