Sys: typedef PetscObjectState and change all accessors
State is used to manage caches of intermediate structures.  If state
overflowed, it would allow a cached value to be reused incorrectly,
leading to wrong results or possibly corruption.  In long runs, it is
conceivable for an object to undergo 2^32 state transitions between
encounters, thus it is important to use a 64-bit counter.  The typedef
is just for clarity.