Original commit by Paul Mullowney (augmented by PETSC_HAVE_CUSP include guards by Karl Rupp): Implementation of Sequential to Sequential VecScatters on the GPU.
In this commit, I've built a working prototype for sequential
to sequential vector scatters for CUSP vectors. I've also reorganized
the parallel to parallel vector scatters in a new infrastructure.

The design of the code is as follows. Currently, I distinguish
between PtoP (parallel to parallel) VecScatters and StoS (sequential
to sequential) VecScatters. In cuspvecimpl.h, a high level struct
called PetscCUSPIndices stores a void * pointer and enumerated type
describing whether the scatter is PtoP or StoS (and later PtoS and
perhaps StoP). The actual type of scatter information built
and stored in the void * pointer depends on the calling code.

For instance, in vscat.c, there are the methods VecScatterBegin_SGToSG,
VecScatterBegin_SGToSS, ... These methods build the indices for
the sequential scatter on demand via the function
VecScatterCUSPIndicesCreate_StoS.  Those indices are stored in
a struct of type _p_VecScatterCUSPIndices_StoS. the appropriate
data in that struct is filled depending on whether the input
and output are general or strided. Some additional meta data is
also stored.

A similar routine is called from vpscat.c for building the
PtoP vector scatter. The structure of this hasn't changed although
the constructing API method is now VecScatterCUSPIndicesCreate_PtoP
(it used to be PetscCUSPIndicesCreate).

A single destructor function is given for the PtoP and StoS scatters:
VecScatterCUSPIndicesDestroy. It used to be PetscCUSPIndicesDestroy.

I've made all of these methods PETSC_INTERN. I can't see any
reason to expose these methods to the user.

All of the source code for doing the sequential scatter computation
on the GPU is moved into a file vecscattercusp.cu (src/vec/vec/impls/seq/seqcusp).
I haven't moved the VecCUSPCopyToGPUSome and similar methods into
this file, although perhaps they belong there.

All scatter types implemented including FORWARD and REVERSE as well
as INSERT, ADD, MAX.