Mat BAIJ: respect MAT_ROW_ORIENTED in Mat*BAIJSetPreallocationCSR
MatMPIBAIJSetPreallocationCSR() used to handle the input matrix entries
in a peculiar way, treating them as if each row of values was provided
ordered by row within the block rows instead of by block as it was
originally intended to be.  Now treat as array of contiguous blocks for
which the ordering within blocks is row-major or column-major according
to the MAT_ROW_ORIENTED option.

This also affects MatCreateMPIBAIJWithArrays() which now expects
contiguous blocks with column-major ordering within blocks, just like
the internal storage format.  MatCreateSeqBAIJWithArrays is unaffected
since it uses the array as provided.

Reported-by: Matteo Parsani <parsani.matteo@gmail.com>
Reported-by: Lisandro Dalcin <dalcinl@gmail.com>
Helped-by: Barry Smith <bsmith@mcs.anl.gov>