KSP ex49: restrict near-null space for elasticity example with sym BC
This example enforces boundary conditions in a non-symmetric way and has
an option to "lift" the contributions to the right hand side.  Formerly,
the near-null space was not transfered to the constrained operator,
causing smoothed aggregation methods to perform poorly.  This commit
restricts the near-null vectors and attaches them to the reduced-space
operator.

Reported-by: Garth N. Wells <gnw20@cam.ac.uk>