PCGAMG: set default type (AGG); PCSetUp_GAMG crashes if not set
If PCSetFromOptions had not been called, PCSetUp_GAMG would crash due to
missing subtype.  It is more conventional with top-level objects to use
a lazy choice of type, but there is no great setup involved, and no
known order-dependence, so it looks safe to simply set the type to the
default value (AGG) up-front.

Reported-by: Einar Sørheim <einar.sorheim@gmail.com>
Based-on-patch-by: Mark Adams <mfadams@lbl.gov>