gmakegen.py: actually fix symlink translation
We need os.path.realpath to resolve a path without symlinks.  The
problem is that relative paths obtained via symlinks are not valid:

$ ln -s $PETSC_DIR /tmp/petsc
$ cd /tmp/petsc
$ python -c "import sys; sys.path.insert(0, 'conf'); import gmakegen, os; print(os.path.relpath('conf/gmakegen.py', '/tmp/petsc'))"
../../home/jed/petsc/conf/gmakegen.py
$ ls ../../home/jed/petsc/conf/gmakegen.py
ls: cannot access ../../home/jed/petsc/conf/gmakegen.py: No such file or directory

So this relative path expects that ../.. will be evaluated like "cd"
instead of like other commands, that is, relative to the symlinked path
in $PWD instead of the real path.  os.path.realpath resolves all
symlinks and fixes the problem (tested on surveyor).