PetscStack: use same definitions for PetscStack functions
If the user compiles their code with debugging, perhaps calling
PetscStack functions, it would be useful for the implementations to make
a best-effort at displaying the stack usefully when they subsequently
LD_PRELOAD an optimized version of the library.  These functions are
tiny so code size is not an excuse for keeping the complexity of
different versions.