TS: fix for TSGetRHSMats_Private() in the creation of the Arhs and Brhs matrices
When an RHS Jacobian is provided, matrices must be copied or duplicated
for Arhs and Brhs.  Commit e1244c694be870577ddfeb27d69a78d09ad24070

    TS: add TSRHSJacobianSetReuse, fixes TSComputeRHSJacobianConstant

changed the logic for how to recompute RHS Jacobians, which had the
side-effect of causing TSGetRHSMats_Private to duplicate Brhs even if A
was being reused and A == B.  This commit fixes the logic to reuse Arhs
as Brhs when they should match.

This commit cherry-picks the change to src/ts/interface/ts.c from
3162be4ab4fd0bc00dad2f51f4d64a4f9f8a384e.

Reported-by: Jay Jay Billings <jayjaybillings@gmail.com>