ctetgen: require OpenMP to ensure that compiler flags are set
ctetgen does not require OpenMP, but it references petscstack, which is
a thread-local variable.  If OpenMP is enabled, petscstack will be
marked with the pragma omp threadprivate(petscstack), which will be
silently ignored if the compiler is not passed -fopenmp.  This leads to
link errors later when we try to include libctetgen.a (missing
threadprivate) into libpetsc.so.