compilerOptions.py: use -O0 with gcc when in debug mode
GCC's -fno-inline is the default when optimization is off (-O0), so we
do not need to pass it again.  Pass the flag to the C++ and Fortran
compilers because that was missed in the original.

  commit e3218094707942d5674f57c17eb63a9af846a945
  added -fno-inline for better debugging with gnu compilers; take it out if it breaks something

This is not perfect because it only applies to GNU and Clang compilers,
but those are the most common used for debugging.