Several essential feature additions and internal modifications for DMMoab.
1) Petscify the DMMoab implementation and separate out the Mat and Vec handling parts to separate files.
2) Private data structure definitions moved to petsc-private.
3) Other feature additions include clean ways to perform local-global and global-local scatters.
4) Now there is a separate SetUp routine that handles consistency checks and initialization all necessary data for DMMoab.
5) The Matrix creation from DMMoab can handle AIJ and BAIJ type matrix creations based on how the user defined a tag. This assumes that
the primary solution is defined as a single tag in MOAB and hence block size is set in DMMoab globally. This should change in order
to allow more flexibility to create block matrices for single solution fields or combinations of them. Similar changes for Vec can
be provided based on tag structure. Perhaps taking in a MoabVec input can resolve this to maintain consistency.