Nightly Summary: Added filtered output such that only warnings and errors are displayed.
In addition to a link to the full log file, this commit also adds
filtered output to the dashboard, only displaying the lines containing
a warning or error.
For each log file $f, an additional log file "filtered-$f" is generated
in the log file folder from a cascade of suitable grep commands.
There should be no changes to the nightly logs upload script required
as long as it first calls runhtml.sh on the respective logs (branch)
and then uploads the log files to the FTP space.