Vec: document when InsertMode cannot change between reassemblies
When using VEC_SUBSET_OFF_PROC_ENTRIES, the header is currently not
re-communicated.  If the receiving process did not set any values
itself, it would have no way to determine which InsertMode should be
used.  The restriction can be lifted by sending the InsertMode at the
start of each integer message, but that would require out-of-place
message packing.