f2cblaslapack: fix quad version of ila*l[rc] i*max1 functions
Fixed-by: Eloy Romero Alcalde <elroal@upvnet.upv.es>

Reported-by: Barry Smith <bsmith@mcs.anl.gov>

      I stumbled across bugs in f2cblaslapack-3.4.2.q/  for quad/__float128 precision that need fixing.

      In the Lapack directory ila*lc.c  ila*lr.c are single/double real/complex versions but the toclapack.sh
(or whatever generates the quad precision files) does not generate the needed ilaqlc.c  ilaqlr.c and ilawlc.c
ilawlr.c files. In addition, calls to iladlc_ and iladlr_ (same for z) in the q/w source files are not
correctly mapped to the q (same for w) versions instead they are left as is (in for example qlarf.c)

    extern integer iladlc_(integer *, integer *, quadreal *, integer *),
            iladlr_(integer *, integer *, quadreal *, integer *);

/*     Scan for the last non-zero column in C(1:lastv,:). */
            lastc = iladlc_(&lastv, n, &c__[c_offset], ldc);
        } else {
/*     Scan for the last non-zero row in C(:,1:lastv). */
            lastc = iladlr_(m, &lastv, &c__[c_offset], ldc);
        }

Thus when compiling qlarf.c the C compiler thinks iladlc_ takes a quadreal when in fact it takes a double and
errors occur at runtime in the call.

The fix would be to have the script generate the q and w versions of ila*lc.c  ila*lr.c  and when generating
the quad q*.c and w*.c files, replace the calls to iladlc_  iladlr_ (and z version) with ilqdlc_  ilqdlr_  (and
z versions)