Mat & Vec: make block sizes lazy
In the past, MatSetUp eagerly set the block sizes (to 1) if not already
set.  This prevents use of the "convenience" constructors because they
call MatSetUp before returning.  We could add bs as an explicit argument
to those constructors, but block size is an extraneous concept for most
users.  The default value of the block size is -1, so now we use
PetscAbs(map->bs) anywhere it is not known in advance that bs has been
explicitly set.  This allows block sizes to be lazy, and they can be set
arbitrarily late, so long as the new size is compatible with the
existing distribution.  For (S)BAIJ matrices, the block size is
structural so it gets set explicitly and can thus be referenced directly
rather than via MatGetBlockSize or wrapped in PetscAbs.