malloc: PETSC_USE_MALLOC_COALESCED, work around IBM XL miscompile
IBM XL 12.1 for BG/Q and POWER at optimization level -O and above
miscompiles code such as the following, as reduced from PetscMalloc3.

  struct {double *a,*b,*c;} *t;
  t = malloc(sizeof *t);
  (*(void**)&t->a = malloc(24)) && (*(void**)&t->b = t->a+1,*(void**)&t->c = t->b+1,0);
  printf("%p %p %p\n",(void*)t->a,(void*)t->b,(void*)t->c);

The compiler incorrectly allows reordering of

  *(void**)&t->b = t->a+1;

as though this statement did not modify t->b.  To work around this, and
to be more flexible on malloc policy (including debugging malloc
issues), we add PETSC_USE_MALLOC_COALESCED which is normally enabled for
debugging=0 builds, but not when using broken compilers.