TSEvent: Avoid re-detection of the same event when previous function value is exactly zero.
If a previous event results in an event function value of exactly zero, then the same event
was being detected in the next interval. This was because the check for event detection was
based only on the change of function sign. Added an extra guard that checks if the previous
function value is zero and subsequently avoids re-detecting of the previous event.
Observed this behavior in time-based events (t - t_event_time = 0).